refactor: ClassPage 컴포넌트 및 SSR 구현 #154
